refactor: 클래스별 구조 분리
