+++ src/main/java/racingcar/domian/RacingGame.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.domian;
+
+import racingcar.ui.OutputUI;
 
 import java.util.ArrayList;
 import java.util.Comparator;
