+++ src/main/java/racingcar/ui/InputUI.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.ui;
 
 import java.util.List;
 import java.util.Scanner;
